requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7.5
django-toolbelt==0.0.1
docutils==0.12
eventbrite==3.0.4
facebook-sdk==0.4.0
foursquare==2015.2.1
google-api-python-client==1.4.0
gunicorn==19.2.1
httplib2==0.9
Jinja2==2.7.3
MarkupSafe==0.23
oauth2client==1.4.7
oauthlib==0.7.2
psycopg2==2.6
pyasn1==0.1.7
pyasn1-modules==0.0.5
Pygments==2.0.2
python-dateutil==2.4.1
python-firebase==1.2
python-instagram==1.3.0
pytz==2015.2
requests==2.6.0
requests-oauthlib==0.4.2
rsa==3.1.4
simplejson==3.6.5
six==1.9.0
Sphinx==1.2.3
static3==0.5.1
twython==3.2.0
uritemplate==0.6
urllib3==1.10.2
